Change of strategy to limit vpd dependence of f_vpd
Instead of having f_vpd (as prposed 2 commits ago), f_vpd is left as variable but the value of the parameter in the related function is set very small (1e-12) in SpeciesParams.tab of the CaseStudy.
This effectively remove dependence, without touching the code, and users can increase the value of f_vpd_coeff if they want to play around with it.